requirements.txt
Django==3.0.2
gunicorn==20.0.4
pipenv==2018.11.26
psycopg2-binary==2.8.4
requests==2.22.0
sqlparse==0.3.0
tablib==0.14.0
urllib3==1.25.7
virtualenv==16.7.9
SQLAlchemy==1.3.12
djangorestframework==3.11.0
